Reduce effect of solid colour backgrounds on similarity
Naturally, two images with a white background will have high similarities, not anymore! It now calculates the similarity between the two background colours, then how much of the images it makes up, then reduces the similarity by that much. The result is that backgrounds now have very little effect on similarity! I'm quite proud of this as it was entirely my desgin :) no copy paste!